refactor(training): relocate standalone modules to canonical packages
